requirements.txt
# Requires Python 3.12
google-genai==1.46.0
numpy==2.3.4
scikit-learn==1.7.2
tqdm==4.67.1
pydantic==2.12.3
hydra-core==1.3.2
omegaconf==2.3.0
markdown==3.9
faiss-cpu==1.12.0
openai>=1.3.0
